requirements.txt
astroid==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysql-connector-python==8.0.17
protobuf==3.9.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
